// citpwawcl/src/app/models/controllo-model.ts
export class ControlloModel {
    idAllegato: number;
    fkStatoRapp: number;
    desStatoRapp: string;
    fkEspezIspet: number;
    fkTipoDocumento: number;
    desTipoDocumento: string;
    fkSiglaBollino: string;
    fkNumeroBollino: number;
    dataControllo: string;
    bFlgLibrettoUso: number;
    bFlgDichiarConform: number;
    bFlgLibImp: number;
    bFlgLibCompldi: number;
    fOsservazioni: string;
    fRaccomandazioni: string;
    fPrescrizioni: string;
    fFlgPuoFunzionare: number;
    fInterventoEntro: string;
    fOraArrivo: string;
    fOraPartenza: string;
    fDenominazTecnico: string;
    fFlgFirmaTecnico: number;
    fFlgFirmaResponsabile: number;
    dataInvio: string;
    dataRespinta: string;
    nomeAllegato: string;
    aPotenzaTermicaNominaleMax: string;
    dataUltMod: string;
    utenteUltMod: string;
    elencoCombustibili: string;
    elencoApparecchiatura: string;
    fkPgCat: number;
    altroDescr: string;
    desRuolo: string;
    ruoloFunz: string;
    idPersonaGiuridica: number;
    pgDenominazione: string;
    pgCodiceFiscale: string;
    pgFkStatoPg: number;
    pgSiglaRea: string;
    codiceImpianto: number;
    comuneImpianto: string;
    siglaProvImpianto: string;
    indirizzoUnitaImmob: string;
    civicoUnitaImmob: string;
    flgControlloBozza: number;
    uidIndex: string;
    idTipoManutenzione: number;
    descTipoManutenzione: string;
    dtInvioMemo: string;
    mailInvioMemo: string;
    uidIndexFirmato: string;
    nomeAllegatoFirmato: string;
}